requirements.txt
astroid==1.6.1
colorama==0.3.9
Django==1.9.1
django-markdownx==2.0.22
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.0.0
pylint==1.8.2
six==1.11.0
wrapt==1.10.11
